% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-vcf-annotation.R
\name{add_rsids_to_funseq}
\alias{add_rsids_to_funseq}
\title{Add rs ID to INFO annotation}
\usage{
add_rsids_to_funseq(vcf_genotypes, info_data, chrn)
}
\arguments{
\item{vcf_genotypes}{CollapsedVCF, genotypes vcf}

\item{info_data}{this can be the output of \code{\link{get_CSQ}} or \code{\link{get_FUNSEQ}}}

\item{chrn}{int or chr, chromosome name}
}
\value{
A data frame
\describe{
\item{varid}{Variant ID}
\item{range_id}{Id for variant in range chrn:start_RF/ALT}
...
}
}
\description{
the vcf file with variant annotation sometimes does not
include the rs IDs and returns another id when I extract the info
this new id has the format: chrn:start_RF/ALT for some variants
}
\examples{
info_funseq <- get_FUNSEQ(test_anno_vcf)
add_rsids_to_funseq(test_vcf, info_funseq, 22)
}
