% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{consequence_summary}
\alias{consequence_summary}
\title{Summarize the Consequence}
\usage{
consequence_summary(geno_anno_mlong)
}
\arguments{
\item{geno_anno_mlong}{Tibble with genotypes and annotation, the output
of \code{\link{merge_genotypes_with_annotation}}}
}
\value{
A data frame, with the following columns
\describe{
\item{individual}{individual id, example: HG00551}
\item{Consequence}{Variant consequence}
\item{n_variants}{How many variants of type Consequence the}
}
}
\description{
For each individual computes a summary of how many variants types (e. g.
intergenic_variant, stop_lost, etc.) carries.
}
\details{
Consider a particular variant and suppose the consequence is stop_gained.
If the individual is 0|0 it wont carrie the variant, if it is 0|1 or 0|1, it
will carry the variance 1 and 2 for 1|1
}
