% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{count_alt_alleles}
\alias{count_alt_alleles}
\title{Counts the number of alternative alleles}
\usage{
count_alt_alleles(gtype)
}
\arguments{
\item{gtype}{The genotype, example, 0|0, 1|0, etc.}
}
\value{
number of alternative alleles
}
\description{
Counts the number of alternative alleles
}
