% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{funq_load_summary}
\alias{funq_load_summary}
\title{How many variants with (FUNSEQ > cut_off_val) do each individual carry?}
\usage{
funq_load_summary(geno_anno_mlong, cut_off_val)
}
\arguments{
\item{geno_anno_mlong}{Tibble with genotypes and annotation, the output
of \code{\link{merge_genotypes_with_annotation}}}
}
\value{
A data frame, with the following columns
\describe{
\item{individual}{individual id, example: HG00551}
\item{funseq_load}{How many variants have a funseqscore > cut_off_val}
}
}
\description{
For each individual computes how many variants with FUNSEQ > cut_off_val
carries.
}
\details{
If the individual is 0|0 we add nothing to funseq_load
If the individual is 1|0 or 0|1 we add 1 to funseq_load
If the individual is 1|1 or 0|1 we add 2 to funseq_load
}
