% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{funseq_load}
\alias{funseq_load}
\title{Counts the number of alternative alleles}
\usage{
funseq_load(gtype, funseqscore)
}
\arguments{
\item{gtype}{The genotype, example, 0|0, 1|0, etc.}

\item{funseqscore}{the funseq score}
}
\value{
a number 0, 1, or 2.
}
\description{
A funseq score greater than 1.5 is considered deleterious. An indicator function
that returns true if the variant is deleterious multiplied by the numer
of alternative alleles in the genotype.
}
