% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{genetic_load_FUNSEQ_and_Consequence_summary}
\alias{genetic_load_FUNSEQ_and_Consequence_summary}
\title{Computes FUNSEQ load and Consequence summary.}
\usage{
genetic_load_FUNSEQ_and_Consequence_summary(
  vcf_genotypes,
  vcf_annotation,
  chr,
  cores = 6,
  cut_off_val = 1.5
)
}
\arguments{
\item{vcf_genotypes}{CollapsedVCF, genotypes vcf}

\item{vcf_annotation}{CollapsedVCF -> \code{VariantAnnotation::readVcf}. The annotaion for
the variants in \code{vcf_genotypes}}

\item{cores}{int, the number of cores used for parallel computation. The
computation is parallelized over the individuals (samples).}
}
\value{
A list with two elements: FS_l (funseq load) and CSQ_s
}
\description{
The funseq load for an individual is defined as: TODO
}
\details{
NOTE: If an individual has non variant sites (i. e. 0|0) it wont be
present in the results table.
}
