% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-vcf-annotation.R
\name{get_CSQ}
\alias{get_CSQ}
\title{Extract Calculated variant consequence}
\usage{
get_CSQ(vcf_annotation)
}
\arguments{
\item{vcf_annotation}{CollapsedVCF, object read with \code{VariantAnnotation::readVcf}}
}
\value{
A data frame
\describe{
\item{varid}{Variant ID}
\item{Consequence}{Calculated variant consequence}
}
}
\description{
get the Calculated variant consequence
\url{https://m.ensembl.org/info/genome/variation/prediction/predicted_data.html}
}
\examples{
get_CSQ(test_anno_vcf)
}
