% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-vcf-annotation.R
\name{get_FUNSEQ}
\alias{get_FUNSEQ}
\title{Extract FUNSEQ score from vcf annotation info}
\usage{
get_FUNSEQ(vcf_annotation)
}
\arguments{
\item{vcf_annotation}{CollapsedVCF, object read with \code{VariantAnnotation::readVcf}}
}
\value{
A data frame
\describe{
\item{varid}{Variant ID}
\item{FUNSEQ}{funseq score}
}
}
\description{
extract the funseq scores in a tidy data frame from 1k variant annottation.
The FunSeq score is an evaluation of the deleterious effect of noncoding variants.
FunSeq score > 1.5 is suggested to be a cutoff for being deleterious variants,
and used here in some analyses
\url{https://www.internationalgenome.org/faq/where-can-i-get-consequence-annotations-1000-genome-variants/}
}
\examples{
get_FUNSEQ(test_anno_vcf)
}
