% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-vcf-annotation.R
\name{get_var_annotation}
\alias{get_var_annotation}
\title{Extract variant annotation}
\usage{
get_var_annotation(vcf_genotypes, vcf_annotation, chrn)
}
\arguments{
\item{vcf_genotypes}{CollapsedVCF, genotypes vcf}

\item{vcf_annotation}{CollapsedVCF, object read with \code{VariantAnnotation::readVcf}}

\item{chrn}{int or chr, chromosome name}
}
\value{
A data frame
\describe{
\item{varid}{Variant ID}
\item{range_id}{Id for variant in range chrn:start_RF/ALT}
\item{Consequence}{Calculated variant consequence}
\item{FUNSEQ}{FUNSEQ score}
}
}
\description{
This function extracts the variant annotation FUNSEQ and Consequence from
the annotation vcf data
}
\examples{
get_var_annotation(test_vcf, test_anno_vcf, 22)
}
