% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-load-gw.R
\name{gl_funseq_consequence_summary}
\alias{gl_funseq_consequence_summary}
\title{TODO}
\usage{
gl_funseq_consequence_summary(
  vcf_file,
  annotation_file,
  chr,
  cores = 4,
  cut_off_val = 1.5
)
}
\arguments{
\item{cores}{integer, number of cores used for parallel processing}
}
\value{

}
\description{
FUNSEQ load? TODO
}
\details{
CONSEQUENCE summary? TODO
}
