% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{merge_genotypes_with_annotation}
\alias{merge_genotypes_with_annotation}
\title{Merge genotypes with annotation}
\usage{
merge_genotypes_with_annotation(genotipos, annotation, h_samples)
}
\arguments{
\item{genotipos}{the tidy genotypes, output of \code{\link{get_genotype}}}

\item{annotation}{table with annotation, the output of \code{\link{get_var_annotation}}}

\item{h_samples}{vector of samples, a subset of the samples}
}
\value{
A data frame, with the following columns
\describe{
\item{varid}{Variant ID}
\item{Consequence}{Variant consequence}
\item{FUNSEQ}{funseq score}
\item{individual}{individual id, example: HG00551}
\item{genotype}{the genotype of the individual, example: 0|1}
}
}
\description{
Merges the genotypes with the annotation (FUNSEQ and Consequences)
and puts the table in long format
}
