% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_anno_vcf}
\alias{test_anno_vcf}
\title{Variant annotation}
\format{
A VCF file with 300 variants and its corresponding annotation
}
\source{
\url{https://github.com/santiago1234/phdpopgene/blob/main/data-raw/vcf_test_data.R}
}
\usage{
test_anno_vcf
}
\description{
the variant annotation for \code{\link{test_vcf}}
}
\keyword{datasets}
