% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_vcf}
\alias{test_vcf}
\title{A subset of vcf file for chromosome 22 (1000 genomes)}
\format{
A VCF file with 300 variants and a subset of samples from America
}
\source{
\url{https://github.com/santiago1234/phdpopgene/blob/main/data-raw/vcf_test_data.R}
}
\usage{
test_vcf
}
\description{
VCF file used for testing function and package development
}
\keyword{datasets}
