% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_consequences}
\alias{variant_consequences}
\title{Ensembl Variation - Calculated variant consequences}
\format{
A data frame with 36 rows and 3 variables:
\describe{
\item{Consequence}{Variant Consequence}
\item{Display_term}{Variant Consequence in a nice format}
\item{IMPACT}{Variant Impact, HIGH means deleterious}
...
}
}
\source{
\url{https://m.ensembl.org/info/genome/variation/prediction/predicted_data.html}
}
\usage{
variant_consequences
}
\description{
Ensembl Variation - Calculated variant consequences
}
\keyword{datasets}
