% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{funseq_load}
\alias{funseq_load}
\title{Counts the number of alternative alleles}
\usage{
funseq_load(alt_alleles, funseqscore)
}
\arguments{
\item{alt_alleles}{int, how many alternative alles are in the genotype?,
for example 0|0 is 0, 1|0 is 1,}

\item{funseqscore}{the funseq score}
}
\value{
a number 0, 1, or 2.
}
\description{
A funseq score greater than 1.5 is considered deleterious. An indicator function
that returns true if the variant is deleterious multiplied by the number
of alternative alleles in the genotype.
}
