% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis-genetic-load.R
\name{genetic_load_FUNSEQ_and_Consequence_summary}
\alias{genetic_load_FUNSEQ_and_Consequence_summary}
\title{Computes FUNSEQ load and Consequence summary.}
\usage{
genetic_load_FUNSEQ_and_Consequence_summary(
  vcf_genotypes,
  vcf_annotation,
  chr,
  cores = 6
)
}
\arguments{
\item{vcf_genotypes}{CollapsedVCF, genotypes vcf}

\item{vcf_annotation}{CollapsedVCF -> \code{VariantAnnotation::readVcf}. The annotaion for
the variants in \code{vcf_genotypes}}

\item{cores}{int, the number of cores used for parallel computation. The
computation is parallelized over the individuals (samples)}
}
\value{
A list with two elements
}
\description{
The funseq load for an individual is defined as: TODO
}
