% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-load-gw.R
\name{gl_funseq_consequence_summary}
\alias{gl_funseq_consequence_summary}
\title{TODO}
\usage{
gl_funseq_consequence_summary(vcf_file, annotation_file, chr, cores = 4)
}
\arguments{
\item{cores}{}
}
\value{

}
\description{
FUNSEQ load? TODO
}
\details{
CONSEQUENCE summary? TODO
}
